ArcPacker.cs: Compress everything using zlib.
The resulting .arc file is now unpackable using arctool.

Note that arctool expects the file to use a zlib wrapper, which has a
two-byte header and an Adler-32 checksum. I should write a wrapper
class for DeflateStream that handles this.

Also, game.arc has been repacked and tested, and it works.